requirements.txt
asgiref==3.8.1
Django==5.0.7
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-gis==1.1
earthengine-api==1.4.0
Markdown==3.6
matplotlib==3.9.2
sqlparse==0.5.0
pillow==10.2.0
pystac==1.11.0
pystac-client==0.8.5
psycopg==3.2.1
psycopg2==2.9.9
pandas==2.3.3
tzdata==2024.1
typing_extensions==4.12.2
numpy==2.1.3
Werkzeug==3.1.3
pyOpenSSL==24.3.0
django-extensions==3.2.3